split current eth_calls decoder into submodules
Break the 1011-line monolith into focused submodules:
- mod.rs: main decode loop and message dispatch
- state.rs: DecoderState struct
- events.rs: event-triggered call handling
- factory.rs: once/factory call handling
- range_processor.rs: regular call and block-complete handling
